Move code to other files in graph construction. (#94)
*Description of changes:*
This is the first step to refactor the code of graph construction.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.
